Block scanner periodically checks new blocks to process
