+++ supervisord.conf
@@ -56,7 +56,7 @@ stopasgroup=true
 
 [program:block_scanner]
 directory = /code
-command = python manage.py rescan_cashtoken_blocks
+command = python manage.py block_scanner
 autorestart=true
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
@@ -67,7 +67,7 @@ stopasgroup=true
 
 # [program:bchn_zmq_listener]
 # directory = /code
-# command = python manage.py bchn_zmq_listener
+# command = python manage.py mempool_listener
 # autorestart=true
 # stdout_logfile=/dev/stdout
 # stdout_logfile_maxbytes=0
